Refactor parser and tests
